+++ src/Modules/Plugin.php
@@ -1,8 +1,10 @@
 <?php
 
-class Plugin extends RubbishThorClone {
-  use manifest_io;
-  use whippet_helpers;
+namespace Dxw\Whippet\Modules;
+
+class Plugin extends \RubbishThorClone {
+  use Helpers\ManifestIo;
+  use Helpers\WhippetHelpers;
 
   public function commands() {
     $this->command('install', 'Deploys the current set of plugins into your project');
@@ -39,7 +41,7 @@ class Plugin extends RubbishThorClone {
     if(!$this->plugins_lock_file) {
 
       foreach($this->plugins_manifest as $dir => $plugin) {
-        $git = new Git("{$this->plugin_dir}/{$dir}");
+        $git = new \Dxw\Whippet\Git\Git("{$this->plugin_dir}/{$dir}");
 
         // Is the repo there already?
         if(!$git->is_repo()) {
@@ -92,12 +94,12 @@ class Plugin extends RubbishThorClone {
       }
 
       // Delete the ones that don't:
-      $gitignore = new Gitignore($this->project_dir);
+      $gitignore = new \Dxw\Whippet\Git\Gitignore($this->project_dir);
       $ignores = $gitignore->get_ignores();
 
       foreach($plugins_to_delete as $dir) {
         echo "[Removing {$dir}]\n";
-        $git = new Git("{$this->plugin_dir}/{$dir}");
+        $git = new \Dxw\Whippet\Git\Git("{$this->plugin_dir}/{$dir}");
         $git->delete_repo();
 
         // remove from ignores:
@@ -119,7 +121,7 @@ class Plugin extends RubbishThorClone {
       //
 
       foreach($this->plugins_locked as $dir => $plugin) {
-        $git = new Git("{$this->plugin_dir}/{$dir}");
+        $git = new \Dxw\Whippet\Git\Git("{$this->plugin_dir}/{$dir}");
 
         if(!$git->is_repo()) {
           // The repo has gone missing. Let's add it back.
@@ -172,7 +174,7 @@ class Plugin extends RubbishThorClone {
 
         echo "[Adding {$dir}] ";
 
-        $git = new Git("{$this->plugin_dir}/{$dir}");
+        $git = new \Dxw\Whippet\Git\Git("{$this->plugin_dir}/{$dir}");
 
         // Is the repo there already?
         if(!$git->is_repo()) {
@@ -206,7 +208,7 @@ class Plugin extends RubbishThorClone {
     //
     // Make sure that Whippet-managed plugins are gitignored
     //
-    $gitignore = new Gitignore($this->project_dir);
+    $gitignore = new \Dxw\Whippet\Git\Gitignore($this->project_dir);
     $ignores = $gitignore->get_ignores();
 
     foreach($this->plugins_locked as $dir => $plugin) {
@@ -243,7 +245,7 @@ class Plugin extends RubbishThorClone {
       //  - It is the plugin they asked for
       //  - They didn't specify a plugin, and this plugin is in the manifest.
       if($dir == $upgrade_plugin || ($upgrade_plugin == '' && isset($this->plugins_manifest->$dir))) {
-        $git = new Git("{$this->plugin_dir}/{$dir}");
+        $git = new \Dxw\Whippet\Git\Git("{$this->plugin_dir}/{$dir}");
 
         // Find the specified revision.
         echo "[Checking {$dir}] ";
